+++ src/mailman/commands/tests/test_cli_unshunt.py
@@ -33,13 +33,70 @@ class TestUnshunt(unittest.TestCase):
 
     def setUp(self):
         self._command = CliRunner()
-        self._queue = config.switchboards['shunt']
+        self._shunt_q = config.switchboards['shunt']
+        self._in_q = config.switchboards['in']
+
+    def test_unshunt_all(self):
+        filebase1 = self._shunt_q.enqueue(Message(), {})
+        filebase2 = self._shunt_q.enqueue(Message(), {})
+        self.assertEqual(len(self._shunt_q.files), 2)
+        self.assertEqual(len(self._in_q.files), 0)
+        result = self._command.invoke(unshunt, ('--verbose'))
+        self.assertEqual(
+            result.output,
+            f'Unshunting to in queue: {filebase1}\n'
+            f'Unshunting to in queue: {filebase2}\n'
+        )
+        self.assertEqual(len(self._shunt_q.files), 0)
+        self.assertEqual(len(self._in_q.files), 2)
+
+    def test_unshunt_specified_filebase(self):
+        filebase1 = self._shunt_q.enqueue(Message(), {})
+        filebase2 = self._shunt_q.enqueue(Message(), {})
+        filebase3 = self._shunt_q.enqueue(Message(), {})
+        self.assertEqual(len(self._shunt_q.files), 3)
+        self.assertEqual(len(self._in_q.files), 0)
+        result = self._command.invoke(unshunt, ('--verbose', filebase3))
+        self.assertEqual(
+            result.output,
+            f'Unshunting to in queue: {filebase3}\n'
+        )
+        self.assertEqual(len(self._shunt_q.files), 2)
+        self.assertEqual(len(self._in_q.files), 1)
+        result = self._command.invoke(
+            unshunt,
+            ('--verbose', filebase1, filebase2)
+        )
+        self.assertEqual(
+            result.output,
+            f'Unshunting to in queue: {filebase1}\n'
+            f'Unshunting to in queue: {filebase2}\n'
+        )
+        self.assertEqual(len(self._shunt_q.files), 0)
+        self.assertEqual(len(self._in_q.files), 3)
 
     def test_dequeue_fails(self):
-        filebase = self._queue.enqueue(Message(), {})
-        with patch.object(self._queue, 'dequeue',
+        filebase = self._shunt_q.enqueue(Message(), {})
+        self.assertEqual(len(self._shunt_q.files), 1)
+        self.assertEqual(len(self._in_q.files), 0)
+        with patch.object(self._shunt_q, 'dequeue',
                           side_effect=RuntimeError('oops!')):
-            results = self._command.invoke(unshunt)
+            result = self._command.invoke(unshunt)
         self.assertEqual(
-            results.output,
-            'Cannot unshunt message {}, skipping:\noops!\n'.format(filebase))
+            result.output,
+            f'Cannot unshunt message {filebase}, skipping: oops!\n'
+        )
+        self.assertEqual(len(self._shunt_q.files), 1)
+        self.assertEqual(len(self._in_q.files), 0)
+
+    def test_discard(self):
+        filebase = self._shunt_q.enqueue(Message(), {})
+        self.assertEqual(len(self._shunt_q.files), 1)
+        self.assertEqual(len(self._in_q.files), 0)
+        result = self._command.invoke(unshunt, ('--verbose', '--discard'))
+        self.assertEqual(
+            result.output,
+            f'Discarding: {filebase}\n'
+        )
+        self.assertEqual(len(self._shunt_q.files), 0)
+        self.assertEqual(len(self._in_q.files), 0)